backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
requests
Werkzeug==3.0.1
